backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
Pillow
psycopg2-binary
